+++ R/delta.R
@@ -3,7 +3,8 @@ function(data,
          model, 
          type = c("response", "link", "terms"), 
          vcov = vcov(model),
-         eps = 1e-7) {
+         eps = 1e-7,
+         ...) {
     # take the derivative of each marginal effect from a model with respect to each model coefficient
     
     type <- match.arg(type)
@@ -20,19 +21,19 @@ function(data,
     # http://www.soderbom.net/lecture10notes.pdf
     # http://stats.stackexchange.com/questions/122066/how-to-use-delta-method-for-standard-errors-of-marginal-effects
     
-    FUN <- .build_grad_fun(data = data, model = model, type = type)
+    FUN <- .build_grad_fun(data = data, model = model, type = type, ...)
     #jacobian <- numDeriv::jacobian(FUN, model[["coefficients"]], method = "simple")
     jacobian <- jacobian(FUN, model[["coefficients"]], eps = eps)
     vout <- diag(jacobian %*% vcov %*% t(jacobian))
     return(vout)
 }
 
-.build_grad_fun <- function(data, model, type = "response", eps = 1e-7) {
+.build_grad_fun <- function(data, model, type = "response", eps = 1e-7, ...) {
     
     # factory function to return prediction holding data constant but varying coefficients
     FUN <- function(coefs) {
         model[["coefficients"]] <- coefs
-        colMeans(marginal_effects(model = model, data = data, type = type, eps = eps), na.rm = TRUE)
+        colMeans(marginal_effects(model = model, data = data, type = type, eps = eps, ...), na.rm = TRUE)
     }
     return(compiler::cmpfun(FUN))
 }