+++ R/get_effect_variances.R
@@ -23,7 +23,7 @@ function(data,
     } else if (vce == "delta") {
         
         # default method
-        variances <- delta_once(data = data, model = model, type = type, vcov = vcov, eps = eps)
+        variances <- delta_once(data = data, model = model, type = type, vcov = vcov, eps = eps, ...)
         
     } else if (vce == "simulation") {
         
@@ -37,7 +37,7 @@ function(data,
         # estimate AME from from each simulated coefficient vector
         effectmat <- apply(coefmat, 1, function(coefrow) {
             tmpmodel[["coefficients"]] <- coefrow
-            colMeans(marginal_effects(data, model = tmpmodel, type = type))
+            colMeans(marginal_effects(data, model = tmpmodel, type = type, ...))
         })
         # calculate the variance of the simulated AMEs
         variances <- apply(effectmat, 1, var, na.rm = TRUE)
@@ -47,7 +47,7 @@ function(data,
         # function to calculate AME for one bootstrap subsample
         bootfun <- function() {
             s <- sample(seq_len(nrow(data)), nrow(data), TRUE)
-            colMeans(marginal_effects(model = model, data = data[s,], type = type), na.rm = TRUE)
+            colMeans(marginal_effects(model = model, data = data[s,], type = type, ...), na.rm = TRUE)
         }
         # bootstrap the data and take the variance of bootstrapped AMEs
         variances <- apply(replicate(iterations, bootfun()), 1, var, na.rm = TRUE)